% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgenerator.r
\name{generators}
\alias{generators}
\alias{\%..\%}
\alias{liota}
\alias{larith}
\alias{lrange}
\alias{lseq_gen1}
\alias{lseq_gen2}
\alias{lseq_gen3}
\alias{lseq_gen}
\title{Lazy stream generators}
\usage{
lhs \%..\% rhs

liota(n = Inf, start_ = 0L, step_ = 1L)

larith(start_, next_, end_ = Inf)

lrange(start_ = 0L, step_ = 1L, end_ = Inf)

lseq_gen1(x0, f)

lseq_gen2(x0, x1, f)

lseq_gen3(x0, x1, x2, f)

lseq_gen(..., f)
}
\arguments{
\item{lhs}{numeric which is converted to integer}

\item{rhs}{numeric which is converted to integer}

\item{n}{number of elements}

\item{start_}{start value}

\item{step_}{step value}

\item{next_}{next value}

\item{end_}{env value}

\item{x0, x1, x2}{initial values for generating function \code{f}}

\item{f}{a function for \code{lseq_gen}}

\item{...}{arguments for \code{f}}
}
\description{
generate a lazy stream
}
\examples{
x <- 1 \%..\% 3
y <- 4 \%..\% 5
z <- x \%++\% y # => llist(1, 2, 3, 4, 5)

liota(5) # => llist(0, 1, 2, 3, 4)
larith(0, 2, 10) # => llist(0, 2, 4, 6, 8, 10)
lseq_gen(0, 1, f = function(x, y) x + y) # => fibonacci sequence
lseq_gen1(list(0, 1), f = function(xs) list(xs[[2]], xs[[1]]+ xs[[2]]))
}
