% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/lcolon.r
\name{\%..\%}
\alias{\%..\%}
\title{Colon Operator}
\usage{
lhs \%..\% rhs
}
\arguments{
\item{rhs:}{numeric which is converted to integer}

\item{lhs:}{numeric which is converted to integer}
}
\description{
generate lazy stream
}
\examples{
x <- 1 \%..\% 3
y <- 4 \%..\% 5
z <- x \%++\% y # => llist(1, 2, 3, 4, 5)
}

