% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/basic-functions.r
\name{\%:\%}
\alias{\%:\%}
\title{Creates a lazy stream}
\usage{
x \%:\% y
}
\arguments{
\item{x}{car part}

\item{y}{cdr part}
}
\value{
lazy cons of \code{x} and \code{y}.
}
\description{
The idea and funcition names mainly comes from Scheme (Gauche) and Haskell. see Details.
}
\details{
\itemize{
  \item \url{http://srfi.schemers.org/srfi-41/srfi-41.html}
  \item \url{http://srfi.schemers.org/srfi-41/srfi-41.html}
  \item \url{http://srfi.schemers.org/srfi-45/srfi-45.html}
  \item \url{http://practical-scheme.net/gauche/man/gauche-refe_59.html}
  \item \url{http://practical-scheme.net/gauche/man/gauche-refe_88.html}
  \item \url{http://hackage.haskell.org/package/base-4.7.0.0/docs/Prelude.html}
  \item \url{http://hackage.haskell.org/package/base-4.7.0.0/docs/Data-List.html}
}
}
\examples{
# R's infix operator has left-associativity
# run
1 \%:\% (2 \%:\% (3 \%:\% NULL))

# can be parsed but not meaningful to interpret
1 \%:\% 2 \%:\% 3 \%:\% NULL

# repeat 1L as an infinit sequence
ones <- 1L \%:\% ones
}

