% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lindexing.r
\name{\%!!\%}
\alias{\%!!\%}
\title{Extract the nth elements (starting from 0)}
\usage{
x \%!!\% n
}
\arguments{
\item{x}{a lazy stream object}

\item{n}{index starting from 0}
}
\description{
Extract the nth elements (starting from 0)
}
\examples{
x <- llist(1, 2, 3, 4, 5)
x \%!!\% 3 # => 4
}

