% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/lappend.r, R/lconcat.r
\name{lappend}
\alias{\%++\%}
\alias{lappend}
\alias{lconcatMap}
\title{append lazy stream}
\usage{
lhs \%++\% rhs

lappend(lhs, rhs)

lconcatMap(f, x)
}
\arguments{
\item{lhs:}{a lazy stream}

\item{rhs:}{a lazy stream}
}
\description{
append lazy stream
}
\examples{
x <- llist(1, 2)
y <- llist(3, 4, 5)
z <- x \%++\% y # => llist(1, 2, 3, 4, 5)
lappend(x, y) # => llist(1, 2, 3, 4, 5)
}

