% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lconcat.r
\name{lconcat}
\alias{lconcat}
\alias{lconcatMap}
\title{Concat a lazy object}
\usage{
lconcat(x)

lconcatMap(x, f)
}
\arguments{
\item{x}{a lazy stream object}

\item{f}{function that returns a lazy stream object}
}
\value{
[[a]] -> [a]
}
\description{
Concat a lazy object
}
\examples{
lconcat(llist(llist(1,2), llist(3), llist(4,5,6))) # => llist(1,2,3,4,5,6)
lconcatMap(function(x) llist(x, 10 * x), llist(1,2,3)) # => llist(1,10,2,20,3,30)
}

