% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcons.r
\name{lcons}
\alias{lcons}
\alias{\%:\%}
\alias{lcons}
\title{Creates a pair, the right value is promised (by making lambda)}
\usage{
lhs \%:\% rhs

lcons(lhs, rhs)
}
\arguments{
\item{lhs}{car part}

\item{rhs}{cdr part}
}
\value{
lazy cons of \code{lhs} and \code{rhs}.
}
\description{
The idea and funcition names mainly comes from Scheme (Gauche) and Haskell. see References
}
\examples{

# R's infix operator has left-associativity
# run
1 \%:\% (2 \%:\% (3 \%:\% lempty))

# can be parsed but not llist
1 \%:\% 2 \%:\% 3 \%:\% lempty

# repeat 1L as an infinit sequence
ones <- 1L \%:\% ones

# this was privous version of lcons that evaluate a cdr part as eager-eval.
`\%:\%` <- lcons <- function(x, y) {
  `class<-`(pairlist(head = x, tail = function() y), "lcons")
}
}
\references{
\itemize{
  \item \url{http://srfi.schemers.org/srfi-1/srfi-1.html}
  \item \url{http://srfi.schemers.org/srfi-41/srfi-41.html}
  \item \url{http://srfi.schemers.org/srfi-45/srfi-45.html}
  \item \url{http://practical-scheme.net/gauche/man/gauche-refe_59.html}
  \item \url{http://practical-scheme.net/gauche/man/gauche-refe_88.html}
  \item \url{http://hackage.haskell.org/package/base-4.7.0.0/docs/Prelude.html}
  \item \url{http://hackage.haskell.org/package/base-4.7.0.0/docs/Data-List.html}
}
}
