% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lhead.r
\name{lhead}
\alias{lhead}
\title{Extract the first element of lazy stream}
\usage{
lhead(x)
}
\arguments{
\item{x:}{lazy stream objects, which must not be `lempty`}
}
\description{
Extract the first element of lazy stream
}
\examples{
x <- llist(1, 2, 3)
lhead(x) # => 1
}

