% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/linfinit.r
\name{linfinit}
\alias{lcycle}
\alias{linfinit}
\alias{literate}
\alias{lrepeat}
\alias{lreplicate}
\title{Infinit lists}
\usage{
literate(x, f)

lrepeat(x)

lreplicate(x, n)

lcycle(x)
}
\arguments{
\item{x:}{a lazy stream object}

\item{f:}{a generating function}
}
\description{
create an infinit lazy stream form a finit lazy stream
}
\examples{
literate(1, function(x) x+1)
lrepeat(1) # => llist(1, 1, 1, ...
lreplicate(1, 3) # => llist(1,1,1)
lcycle(llist(1))
}

