% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/linit.r
\name{linit}
\alias{linit}
\title{Extract a lazy stream except the last element.}
\usage{
linit(x)
}
\arguments{
\item{x:}{lazy stream objects, which must not be `lempty`}
}
\description{
Extract a lazy stream except the last element.
}
\examples{
x <- llist(1, 2, 3)
linit(x) # => llist(1, 2)
}

