% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/llast.r
\name{llast}
\alias{llast}
\title{Extract the last element of a lazy stream}
\usage{
llast(x)
}
\arguments{
\item{x:}{lazy stream objects, which must not be `lempty`}
}
\description{
Extract the last element of a lazy stream
}
\examples{
x <- llist(1, 2, 3)
llast(x) # => 3
}

