% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/llist.r
\name{llist}
\alias{as.llist}
\alias{llist}
\alias{llist_lazy}
\title{Make a finite lazy object from R's object}
\usage{
llist(...)

llist_lazy(...)

as.llist(x)
}
\arguments{
\item{...}{non lazy-stream object,}

\item{x}{any type of object}
}
\description{
functions to construct and coerce lazy stream
}
\examples{
x <- llist(1, 2, 3) # arguments are evaluated
# y <- llist(1, repeat() {}, 3) # cannot return
z <- llist_lazy(1, repeat{}, 3) # can be defined
lhead(z) # => 1

# Convert R's vector (include R's list) to llist
# is.llist(as.llist(x)) is ALWAYS TRUE
as.llist(1:3) # => llist(1,2,3)
}

