% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/lnull.r
\name{lnull}
\alias{lnull}
\title{Check weather a lazy stream is lempty}
\usage{
lnull(x)
}
\arguments{
\item{x:}{a lazy stream object}
}
\description{
Check weather a lazy stream is lempty
}
\details{
'lempty' is just a symbol representing empty list of a lazy stream object.
 lnull() tests whether a R object equals lempty.
}
\examples{
lnull(llist(1, 2, 3)) # => FALSE
lnull(lempty) # => TRUE
}

