% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsearch.r
\name{lsearching}
\alias{lsearching}
\alias{lfind}
\alias{lfilter}
\alias{lpartition}
\title{Searching with predicate f}
\usage{
lfind(x, f)

lfilter(x, f)

lpartition(x, f)
}
\arguments{
\item{x}{a lazy stream}

\item{f}{a predicate function (a' -> boolean)}
}
\description{
Searching with predicate f
}
\examples{
lfilter(1\%..\%10, function(x) x\%\%3== 0) # => llist(3,6,9)

}
