% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ltail.r
\name{ltail}
\alias{ltail}
\title{Extract the rest of the first element of lazy stream}
\usage{
ltail(x)
}
\arguments{
\item{x:}{lazy stream objects, which must not be `lempty`}
}
\description{
Extract the rest of the first element of lazy stream
}
\examples{
x <- llist(1, 2, 3)
ltail(x) # => llist(2, 3)
}

