% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ltake.r
\name{ltake}
\alias{ltake}
\title{Extracting subset of lazy stream}
\usage{
ltake(n, x)
}
\arguments{
\item{n:}{integer}

\item{x:}{a lazy stream}
}
\description{
take the first nth elements from a lazy stream
}
\examples{
x <- 1 \%..\% 10
ltake(5, x) # => llist(1, 2, 3, 4, 5)
ldrop(7, x) # => llist(8, 9, 10)
}

