% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lzip.r
\name{lzip}
\alias{lunzip}
\alias{lunzip2}
\alias{lzip}
\alias{lzipWith}
\title{Zipping and unzipping a lazy stream}
\usage{
lzipWith(f, ...)

lzip(...)

lunzip2(x)

lunzip(x)
}
\arguments{
\item{f}{zipping function}

\item{...}{lazy stream objects}

\item{x}{llist of llist}
}
\description{
When zipping, if the size of arguments are different,
 the shortest one is used for output, i.e., excess elements are discarded.
}
\examples{
lint <- liota()
ltake(lzipWith(`*`, lint, lint), 10)
ltake(lzip(lint, lint, lint), 10)

lunzip(lzip(0\%..\%2, 100\%..\%200, 10\%..\%Inf))
# => list(list(0, 1, 2), list(100, 101, 102), list(10, 11, 12))
}

