% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsublists.r
\name{sublists}
\alias{sublists}
\alias{ltake}
\alias{ldrop}
\alias{lsplitAt}
\alias{ltakeWhile}
\alias{ldropWhile}
\alias{lspan}
\alias{lbreak}
\title{Extracting subset of lazy stream}
\usage{
ltake(x, n)

ldrop(x, n)

lsplitAt(x, n)

ltakeWhile(x, f)

ldropWhile(x, f)

lspan(x, f)

lbreak(x, f)
}
\arguments{
\item{x}{a lazy stream}

\item{n}{intege}

\item{f}{predicate function (a' -> Boolean)}
}
\description{
the order of parameters is different from Haskell.
 The first argument is a lazy steram which enable you to use magrittr's pipe smoothly.
}
\examples{
x <- 1 \%..\% 10
ltake(x, 5) # => llist(1, 2, 3, 4, 5)
ldrop(x, 7) # => llist(8, 9, 10)
lsplitAt(x, 4) # => list(1\%..\%3, 4\%..\%10)
ltakeWhile(x, function(x) x < 4) # => 1\%..\%3
ldropWhile(x, function(x) x < 4) # => 4\%..\%10
lspan(x, function(x) x < 4) # => list(1\%..\%3, 4\%..\%10)
lbreak(x, function(x) x >= 4) # => list(1\%..\%3, 4\%..\%10)
}
