% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/is-functions.r
\name{is_lazystream}
\alias{is.lcons}
\alias{is.llist}
\alias{is.lpair}
\alias{is_lazystream}
\title{is functions}
\usage{
is.lcons(x)

is.lpair(x)

is.llist(x)
}
\arguments{
\item{x}{a lcons object}
}
\description{
is functions
}
\details{
is.llist(x) is TRUE if \code{x} is lempty or ltail(x) is llist.
 is.ilist(x) is TRUE for infinit-stream if ltail(x) and ltail(ltail(x))
  is the same object (circular reference) or the same function except its environment.
}
\examples{
is.lcons(lempty) # => FALSE
 is.lpair(lempty) # => FALSE
 is.llist(lempty) # => TRUE

 is.lcons(1 \%:\% 2) # => TRUE
 is.lpair(1 \%:\% 2) # => TRUE
 is.llist(1 \%:\% 2) # => FALSE

 is.lcons(1 \%..\% 2) # => TRUE
 is.lpair(1 \%..\% 2) # => FALSE
 is.llist(1 \%..\% 2) # => TRUE

 ones <- 1 \%:\% ones
 is.llist(ones) # => TRUE # circular reference
 is.llist(1 \%..\% Inf) #=> TRUE # the same function
}

