% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/is-functions.r
\name{is_lazystream}
\alias{is.llist}
\alias{is.llist_atomic_maybe}
\alias{is.lpair}
\alias{is.lpair_not_llist}
\alias{is_lazystream}
\title{is functions}
\usage{
is.lpair(x)

is.lpair_not_llist(x)

is.llist(x)

is.llist_atomic_maybe(xs, depth = 10)
}
\arguments{
\item{x}{a lcons object}
}
\description{
is functions
}
\details{
is.llist(x) is TRUE if \code{x} is lempty or ltail(x) is llist.
 is.ilist(x) is TRUE for infinit-stream if ltail(x) and ltail(ltail(x))
  is the same object (circular reference) or the same function except its environment.
}
\examples{
is.lpair(lempty) # => FALSE
 is.lpair_not_llist(lempty) # => FALSE
 is.llist(lempty) # => TRUE

 is.lpair(1 \%:\% 2) # => TRUE
 is.lpair_not_llist(1 \%:\% 2) # => TRUE
 is.llist(1 \%:\% 2) # => FALSE

 is.lpair(1 \%..\% 2) # => TRUE
 is.lpair_not_llist(1 \%..\% 2) # => FALSE
 is.llist(1 \%..\% 2) # => TRUE

 ones <- 1 \%:\% ones
 is.llist(ones) # => TRUE # circular reference
 is.llist(1 \%..\% Inf) #=> TRUE # the same function

 is.llist_atomic_maybe(1 \%..\% 10) # => TRUE
 is.llist_atomic_maybe(1 \%..\% 11) # => structure(TRUE, class = "maybe")
 is.llist_atomic_maybe(1 \%..\% Inf) # => structure(TRUE, class = "maybe")
}

