% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lspecail-folds.r
\name{lboolean}
\alias{lall}
\alias{land}
\alias{lany}
\alias{lboolean}
\alias{lor}
\title{Boolean functions}
\usage{
lall(x, f)

lany(x, f)

land(x)

lor(x)
}
\arguments{
\item{x}{a lazy stream object}

\item{f}{a predicate function}
}
\description{
Return TRUE or FALSE for a lazy stream
}
\examples{
x <- 1\%..\%10
lall(x, function(x) x <= 10)
lany(x, function(x) x\%\%7==0)
}

