% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/llength.r
\name{llength}
\alias{llength}
\title{Count length of finite lazy stream}
\usage{
llength(x)
}
\arguments{
\item{x}{a lazy stream}
}
\description{
Count length of finite lazy stream
}
\examples{
x <- llist(1, 2, 3)
llength(x) # => 3
}

