% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/llist-transfomer.r
\name{llist_transformer}
\alias{llist_transformer}
\alias{lmap}
\alias{lmap2}
\alias{lreverse}
\alias{ltranspose}
\title{llist transformation}
\usage{
lmap(x, f)

lmap2(x, ex)

lreverse(x, acc = lempty)

ltranspose(l)
}
\arguments{
\item{x}{a lazy stream}

\item{f}{conver}

\item{ex}{a simbol `.` is treated as an formal parameter}

\item{l}{llist of llist}
}
\description{
llist transformation
}
\examples{
x1 <- llist(llist(1,2,3), llist(4,5,6))
x2 <- ltranspose(x1)
lprint(x1); lprint(x2)

lmap(liota(), function(x) x^2)
lmap2(liota(), .^2)

lreverse(liota(10))
}

