% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/lsublists.r
\name{sublists}
\alias{lbreak}
\alias{ldrop}
\alias{ldropWhile}
\alias{lspan}
\alias{lsplitAt}
\alias{ltake}
\alias{ltakeWhile}
\alias{sublists}
\title{Extracting subset of lazy stream}
\usage{
ltake(x, n)

ldrop(x, n)

lsplitAt(x, n)

ltakeWhile(x, f)

ldropWhile(x, f)

lspan(x, f)

lbreak(x, f)
}
\arguments{
\item{x:}{a lazy stream}

\item{n:}{intege}

\item{f:}{predicate function (a' -> Boolean)}
}
\description{
the order of parameters is different from Haskell.
 The first argument is a lazy steram which enable you to use magrittr's pipe smoothly.
}
\examples{
x <- 1 \%..\% 10
ltake(x, 5) # => llist(1, 2, 3, 4, 5)
ldrop(x, 7) # => llist(8, 9, 10)
lsplitAt(x, 4) # => list(llist(1,2,3,4), llist(5,6,7,8,9,10))
ltakeWhile(x, function(x) x < 4) # => llist(1,2,3)
ldropWhile(x, function(x) x < 4) # => llist(1,2,3)
lspan(x, function(x) x < 4) # => list(1\%..\%3, 4\%..\%10)
lbreak(x, function(x) x >= 4) # => list(1\%..\%3, 4\%..\%10)
}

